prov/efa: Fix wait_send procedure
This leverages a specialized procedure for processing outstanding
completions of an EP to be closed. Unless RM is enabled (FI_RM_ENABLED)
and the packet types are RECEIPT or EOR, the packet entries are simply
released. Otherwise, the packets are queued for reposting if RNR was
encountered. This avoids a lot of overhead with the standard completion
polling procedure.

Signed-off-by: Darryl Abbate <drl@amazon.com>